[HttpKernel] Bundles to [Bundle]
