+++ src/Symfony/Component/Bundle/Tests/Fixtures/ExtensionAbsentBundle/ExtensionAbsentBundle.php
@@ -1,0 +1,18 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Bundle\Tests\Fixtures\ExtensionAbsentBundle;
+
+use Symfony\Component\Bundle\CommandBundle;
+
+class ExtensionAbsentBundle extends CommandBundle
+{
+}
